modules/02_qwen_image_edit_service/requirements_gradio.txt
# Minimal interactive UI
gradio>=4.44
pillow>=10.0
requests>=2.31
huggingface_hub>=0.23

# Qwen-Image-Edit pipeline
# Install diffusers from git per official guidance:
#   pip install git+https://github.com/huggingface/diffusers
